Revert "Feature/update mainframe"
