+++ Core.Backend.Secure/Auth/IDToken.cs
@@ -1,5 +1,4 @@
 using System.Security.Claims;
-using System.Text.RegularExpressions;
 using Newtonsoft.Json;
 
 namespace Core.Backend.Secure.Auth;
@@ -40,7 +39,6 @@ public class IDToken : AuthToken
     {
         get
         {
-
             var claims = new List<Claim>
             {
                 new Claim("type", "id-token"),
@@ -57,16 +55,7 @@ public class IDToken : AuthToken
             if (Class != null)
                 claims.Add(new Claim("klasse", Class));
 
-            var rg = new Regex(@"5[^b]");
-            if (Class != null && rg.Match(Class).Success)
-                claims.Add(new Claim("bessereKlasse", "5b"));
-
             return claims;
         }
-    }
-
-    public enum AvailablePlatforms
-    {
-        Webuntis, Eduvidual, LDAP, Sokrates, None
     }
 }
