+++ Core.Backend.Secure/Services/CredService.cs
@@ -1,6 +1,5 @@
 using System.Security.Cryptography;
 using System.Text;
-using Core.Backend.Secure.Dtos;
 using Core.Backend.Secure.exceptions;
 using Core.Database;
 using Microsoft.EntityFrameworkCore;
@@ -46,11 +45,11 @@ public class CredService
         }
     }
 
-    public void SaveLdapPassword(Guid uuid, LdapUserDto user, bool encrypt = false)
+    public void SaveLdapPassword(Guid uuid, string username, string password)
     {
         var userTokens = GetUserTokenSet(uuid);
-        userTokens.LdapUsername = user.LdapUsername;
-        userTokens.LdapPassword = encrypt ? EncryptPw(user.LdapPassword) : user.LdapPassword;
+        userTokens.LdapUsername = username;
+        userTokens.LdapPassword = password;
         _db.SaveChanges();
     }
 }
