+++ frontend/libs/mainframe-connector/src/lib/mainframe-id-token.service.ts
@@ -9,21 +9,8 @@ export class MainframeIdTokenService {
   constructor() {
   }
 
-  public getJwt(): Promise<string> {
-    parent.postMessage({
-      method: "getIDToken"
-    }, "*");
-
-    const promise = new Promise<string>((resolve) => {
-      window.addEventListener("message", (event) => {
-        if (event.data.method === "getIDToken") {
-          resolve(event.data.data);
-        }
-      });
-    });
-
-    return promise;
-    //return "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0eXBlIjoiaWQtdG9rZW4iLCJ1c2VybmFtZSI6InNmZWljaHRsYmF1ZXIxOCIsInV1aWQiOiJzdXBhIHV1aWQgaG9zdCBkdSBkbyIsInJvbGxlIjoic3R1ZGVudCIsImVtYWlsIjoic2ZlaWNodGxiYXVlcjE4QHN1cy5odGwtZ3JpZXNraXJjaGVuLmF0IiwiY29ubmVjdGVkUGxhdGZvcm1zIjoiW1wiZmVpY2h0bGJhdWVyLmRkbnMubmV0XCIsXCJzcGllbGFmZmUuZGVcIixcInNjaG51cGZ0YWJha2JpbGxpZ2VyMjQuYXRcIl0iLCJtYXRyaWtlbG51bW1lciI6IjY5Iiwia2xhc3NlIjoiNUIifQ._VJ_BC8x6GTTMbQ_ZBKD84UArUWdaSmGOTKj8SghEFM";
+  public getJwt(): string {
+    return "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ0eXBlIjoiaWQtdG9rZW4iLCJ1c2VybmFtZSI6InNmZWljaHRsYmF1ZXIxOCIsInV1aWQiOiJzdXBhIHV1aWQgaG9zdCBkdSBkbyIsInJvbGxlIjoic3R1ZGVudCIsImVtYWlsIjoic2ZlaWNodGxiYXVlcjE4QHN1cy5odGwtZ3JpZXNraXJjaGVuLmF0IiwiY29ubmVjdGVkUGxhdGZvcm1zIjoiW1wiZmVpY2h0bGJhdWVyLmRkbnMubmV0XCIsXCJzcGllbGFmZmUuZGVcIixcInNjaG51cGZ0YWJha2JpbGxpZ2VyMjQuYXRcIl0iLCJtYXRyaWtlbG51bW1lciI6IjY5Iiwia2xhc3NlIjoiNUIifQ._VJ_BC8x6GTTMbQ_ZBKD84UArUWdaSmGOTKj8SghEFM";
   }
 
   public decodeJwt(token: string): IDTokenPayload {
