+++ mainframe/index.html
@@ -52,9 +52,9 @@
     }
 
     function createPluginFrame(url, targetDivId) {
-        const newFrame = document.createElement("iframe");
-        newFrame.src = url;
-        document.getElementById(targetDivId).appendChild(newFrame);
+        const loginFrame = document.createElement("iframe");
+        loginFrame.src = url;
+        document.getElementById(targetDivId).appendChild(loginFrame);
     }
 
     function clearFrameContainer(targetDivId) {
@@ -68,6 +68,7 @@
         return new Promise(resolve => {
             const frame = document.getElementById(containerId).firstChild;
             window.addEventListener("message", function (event) {
+                console.log(event);
                 if (event.source !== frame.contentWindow) {
                     console.debug("Ignoring message from unknown source", event);
                     return;
@@ -101,11 +102,6 @@
         frame.contentWindow.postMessage({method: "error", data: {message}}, "*");
     }
 
-    function sendMessageToFrame(container, method, message) {
-        const frame = document.getElementById(container).firstChild;
-        frame.contentWindow.postMessage({method: method, data: message}, "*");
-    }
-
     async function executeLogin(secureBackendUrl, username, password) {
         const response = await fetch(secureBackendUrl + "/Auth/Login", {
             method: "POST",
@@ -122,28 +118,6 @@
 
         if (response.status === 200)
             return true;
-
-        const error = await response.json();
-
-        if (typeof error === "string")
-            throw new Error(error);
-
-        throw new Error("Es ist ein unbekannter Fehler aufgetreten!");
-    }
-
-    async function fetch2(secureBackendUrl, requestPath,method, body = undefined) {
-        const response = await fetch(secureBackendUrl + requestPath, {
-            method: method,
-            headers: {
-                "Accept": "application/json",
-                "Content-Type": "application/json"
-            },
-            credentials: "include",
-            body: body === undefined ? undefined : JSON.stringify(body)
-        });
-
-        if (response.status === 200)
-            return await response.json();
 
         const error = await response.json();
 
@@ -175,25 +149,13 @@
 
         createPluginFrame(config.navigation.url, "nav-container");
 
-        while (isAuthed) {
-            console.log     ("Waiting for plugin request");
-            const navData = await waitForMessageFromFrame("nav-container", (data) => typeof data.method === "string");
-
-            /*
-             typeof data.method === "string" &&
-                typeof data.pluginId === "string"
-             */
-            /*switch (navData.pluginId) {
-                case "navigate": doNavigateStuff(navData); break;
-            }*/
-
-            switch (navData.method) {
-                case "getIDToken":
-                    sendMessageToFrame("nav-container", "getIDToken", await fetch2(config.secureBackendUrl, "/Auth/GetIDToken", "GET"));
-            }
-
+        while (true) {
+            const navData = await waitForMessageFromFrame("nav-container", (data) =>
+                typeof data.method === "string" &&
+                typeof data.pluginId === "string" &&
+                data.method === "loadPlugin");
             clearFrameContainer("plugin-container");
-            //createPluginFrame(config[navData.pluginId].url, "plugin-container");
+            createPluginFrame(config[navData.pluginId].url, "plugin-container");
         }
     }
